Refactor code: move input choice functionality to main.rb
